pyproject.toml
[project]
name = "oai-ws"
version = "0.1.0"
description = "WebSocket-based Power BI editing agent CLI foundation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.1",
    "websocket-client>=1.9.0",
]

[project.scripts]
pbi-agent = "pbi_agent.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pbi_agent"]
